renamed auth.ts to authentication.ts
