+++ backend/src/lib/authentication.ts
